+++ rio_tiler/io/rasterio.py
@@ -638,6 +638,7 @@ class ImageReader(Reader):
         indexes: Optional[Indexes] = None,
         expression: Optional[str] = None,
         force_binary_mask: bool = True,
+        out_dtype: Optional[Union[str, numpy.dtype]] = None,
         resampling_method: RIOResampling = "nearest",
         unscale: bool = False,
         post_process: Optional[
@@ -675,6 +676,7 @@ class ImageReader(Reader):
             indexes=indexes,
             expression=expression,
             force_binary_mask=force_binary_mask,
+            out_dtype=out_dtype,
             resampling_method=resampling_method,
             unscale=unscale,
             post_process=post_process,
@@ -689,6 +691,7 @@ class ImageReader(Reader):
         height: Optional[int] = None,
         width: Optional[int] = None,
         force_binary_mask: bool = True,
+        out_dtype: Optional[Union[str, numpy.dtype]] = None,
         resampling_method: RIOResampling = "nearest",
         unscale: bool = False,
         post_process: Optional[
@@ -731,6 +734,7 @@ class ImageReader(Reader):
             height=height,
             indexes=indexes,
             force_binary_mask=force_binary_mask,
+            out_dtype=out_dtype,
             resampling_method=resampling_method,
             unscale=unscale,
             post_process=post_process,
@@ -748,6 +752,8 @@ class ImageReader(Reader):
         y: float,
         indexes: Optional[Indexes] = None,
         expression: Optional[str] = None,
+        out_dtype: Optional[Union[str, numpy.dtype]] = None,
+        resampling_method: RIOResampling = "nearest",
         unscale: bool = False,
         post_process: Optional[
             Callable[[numpy.ma.MaskedArray], numpy.ma.MaskedArray]
@@ -760,6 +766,7 @@ class ImageReader(Reader):
             y (float): Y coordinate.
             indexes (sequence of int or int, optional): Band indexes.
             expression (str, optional): rio-tiler expression (e.g. b1/b2+b3).
+            resampling_method (RIOResampling, optional): RasterIO resampling algorithm. Defaults to `nearest`.
             unscale (bool, optional): Apply 'scales' and 'offsets' on output data value. Defaults to `False`.
             post_process (callable, optional): Function to apply on output data and mask values.
 
@@ -775,6 +782,8 @@ class ImageReader(Reader):
             expression=expression,
             unscale=unscale,
             post_process=post_process,
+            out_dtype=out_dtype,
+            resampling_method=resampling_method,
             window=Window(col_off=x, row_off=y, width=1, height=1),
         )
 
@@ -784,6 +793,7 @@ class ImageReader(Reader):
             coordinates=self.dataset.xy(x, y),
             crs=self.dataset.crs,
             band_names=img.band_names,
+            pixel_location=(x, y),
         )
 
     def feature(  # type: ignore
@@ -795,6 +805,7 @@ class ImageReader(Reader):
         height: Optional[int] = None,
         width: Optional[int] = None,
         force_binary_mask: bool = True,
+        out_dtype: Optional[Union[str, numpy.dtype]] = None,
         resampling_method: RIOResampling = "nearest",
         unscale: bool = False,
         post_process: Optional[
@@ -814,6 +825,7 @@ class ImageReader(Reader):
             height=height,
             width=width,
             force_binary_mask=force_binary_mask,
+            out_dtype=out_dtype,
             resampling_method=resampling_method,
             unscale=unscale,
             post_process=post_process,